+++ LLDebugToolSwift/LLog.swift
@@ -37,7 +37,7 @@ public class LLog: NSObject {
     /// Log a normal message with event.
     public static func log(message : String? , event : String? , file : String = #file , function : String = #function , lineNumber : Int = #line) {
         #if DEBUG
-            LLDebugTool.shared().log(inFile: (file as NSString).lastPathComponent, function: function, lineNo: lineNumber, level: .default, onEvent: event, message: message)
+            LLDebugTool.shared().log(inFile: (file as NSString).lastPathComponent, function: function, lineNo: lineNumber, onEvent: event, message: message)
         #endif
     }
     
@@ -49,7 +49,7 @@ public class LLog: NSObject {
     /// Log a alert message with event.
     public static func alertLog(message : String? , event : String? , file : String = #file , function : String = #function , lineNumber : Int = #line) {
         #if DEBUG
-            LLDebugTool.shared().log(inFile: (file as NSString).lastPathComponent, function: function, lineNo: lineNumber, level: .alert, onEvent: event, message: message)
+            LLDebugTool.shared().alertLog(inFile: (file as NSString).lastPathComponent, function: function, lineNo: lineNumber, onEvent: event, message: message)
         #endif
     }
     
@@ -61,7 +61,7 @@ public class LLog: NSObject {
     /// Log a warning message with event.
     public static func warningLog(message : String? , event : String? , file : String = #file , function : String = #function , lineNumber : Int = #line) {
         #if DEBUG
-            LLDebugTool.shared().log(inFile: (file as NSString).lastPathComponent, function: function, lineNo: lineNumber, level: .warning, onEvent: event, message: message)
+            LLDebugTool.shared().warningLog(inFile: (file as NSString).lastPathComponent, function: function, lineNo: lineNumber, onEvent: event, message: message)
         #endif
     }
     
@@ -73,7 +73,7 @@ public class LLog: NSObject {
     /// Log a error message with event.
     public static func errorLog(message : String? , event : String? , file : String = #file , function : String = #function , lineNumber : Int = #line) {
         #if DEBUG
-            LLDebugTool.shared().log(inFile: (file as NSString).lastPathComponent, function: function, lineNo: lineNumber, level: .error, onEvent: event, message: message)
+            LLDebugTool.shared().errorLog(inFile: (file as NSString).lastPathComponent, function: function, lineNo: lineNumber, onEvent: event, message: message)
         #endif
     }